ga_analytics_agent/requirements.txt
# Core Pydantic AI and OpenAI
pydantic-ai>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0

# HTTP and async utilities
httpx>=0.25.0
aiofiles>=23.0.0

# Chart generation for conversational responses
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0
pandas>=2.1.0

# Optional caching
redis>=5.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Web framework for API endpoints
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0.0